Seperated the source filtering in box generator into a seperate image processor
