+++ pyobs/images/processors/misc/__init__.py
@@ -10,3 +10,4 @@ from .removebackground import RemoveBackground
 from .smooth import Smooth
 from .softbin import SoftBin
 from .broadcast import Broadcast
+from .image_source_filter import ImageSourceFilter
