+++ tests/images/processors/misc/test_source_catalog.py
